requirements.txt
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
Flask==2.0.1
Flask_JWT_Extended==4.2.3
email_validator==1.1.3
Flask_PyMongo==2.3.0
pymongo==3.12.0
python_jose==3.3.0
jose==1.0.0
python-decouple==3.4
